% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votesinfo.R
\name{vott vote informaton}
\alias{vott vote informaton}
\alias{vott}
\title{vote information of parliament members}
\usage{
vott (name1,start,end)
}
\arguments{
\item{name1}{A char.name of the politician}

\item{start}{a integer. start year}

\item{end}{a integer. end year}
}
\value{
l. vote information of certain member
}
\description{
get the data from parliament API and return the attendance rate, attitude and participation of each member.
}
\examples{
l1 <- vott("C",2015,2020)
}
