% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votesinfo.R
\name{vott vote informaton}
\alias{vott vote informaton}
\alias{vott}
\title{vote information of parliament members}
\usage{
vott(name1)
}
\value{
vote information of certain member
}
\description{
get the data from parliament API and return the attendance rate, attitude and participation of each member.
}
\examples{
x <- vott()
cal_info(x,A Eriksson, Erik  (C))
}
